// backend/src/shared/types/shared.enum.ts
export enum EventEmitterName {
  AppInstall = 'app.installed',
  ShopUpdate = 'shop.update',
  AppUninstall = 'app.uninstalled',
  PlanPurchase = 'plan.purchase',
  ReportUpdate = 'report.update',
  DataRemove = 'data.remove',
  MetafieldAppUpdate = 'metafield.app.update',
}

export enum SortDirection {
  ASC = 'ASC',
  DESC = 'DESC',
}